requirements.txt
django==3.1.1
djangorestframework==3.11.1
django-extensions
psycopg2
django-filter~=2.3.0